Created MVVM Packages, Moved MainActivity.kt to ui Package and Created MentorManagerApplication.kt app entry point class.
