+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.peculiaruc.alc_mmsystem_mentormanager">
 
     <application
+        android:name=".MentorManagerApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -14,7 +15,7 @@
         android:theme="@style/Theme.AlcMMSystemmentorManager"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
