+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentormanager/MentorManagerApplication.kt
@@ -1,0 +1,5 @@
+package com.peculiaruc.alc_mmsystem_mentormanager
+
+import android.app.Application
+
+class MentorManagerApplication : Application()
