+++ app/src/main/java/com/peculiaruc/alc_mmsystem_mentormanager/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.peculiaruc.alc_mmsystem_mentormanager
+package com.peculiaruc.alc_mmsystem_mentormanager.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.peculiaruc.alc_mmsystem_mentormanager.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
